+++ sprockets/mixins/mediatype/transcoders.py
@@ -8,9 +8,13 @@ Bundled media type transcoders.
 """
 from __future__ import annotations
 
+import array
 import base64
 import dataclasses
+import decimal
+import ipaddress
 import json
+import pathlib
 import string
 import typing
 import urllib.parse
@@ -85,7 +89,8 @@ class JSONTranscoder(handlers.TextContentHandler):
         return typing.cast(type_info.Deserialized,
                            json.loads(str_repr, **self.load_options))
 
-    def dump_object(self, obj: type_info.Serializable) -> str:
+    def dump_object(self,
+                    obj: type_info.Serializable) -> type_info.JsonDumpable:
         """
         Called to encode unrecognized object.
 
@@ -97,28 +102,47 @@ class JSONTranscoder(handlers.TextContentHandler):
         to :func:`json.dumps`.  It provides default representations for
         a number of Python language/standard library types.
 
-        +----------------------------+---------------------------------------+
-        | Python Type                | String Format                         |
-        +----------------------------+---------------------------------------+
-        | :class:`bytes`,            | Base64 encoded string.                |
-        | :class:`bytearray`,        |                                       |
-        | :class:`memoryview`        |                                       |
-        +----------------------------+---------------------------------------+
-        | :class:`datetime.datetime` | ISO8601 formatted timestamp in the    |
-        |                            | extended format including separators, |
-        |                            | milliseconds, and the timezone        |
-        |                            | designator.                           |
-        +----------------------------+---------------------------------------+
-        | :class:`uuid.UUID`         | Same as ``str(value)``                |
-        +----------------------------+---------------------------------------+
+        +--------------------------------+------------------------------------+
+        | Python Type                    | String Format                      |
+        +--------------------------------+------------------------------------+
+        | :class:`bytes`,                | Base64 encoded string.             |
+        | :class:`bytearray`,            |                                    |
+        | :class:`memoryview`            |                                    |
+        +--------------------------------+------------------------------------+
+        | :class:`datetime.datetime`     | ISO8601 formatted timestamp in the |
+        |                                | extended format including          |
+        |                                | separators, milliseconds, and the  |
+        |                                | timezone designator.               |
+        +--------------------------------+------------------------------------+
+        | :class:`uuid.UUID`             | Same as ``str(value)``             |
+        +--------------------------------+------------------------------------+
+        | :class:`decimal.Decimal`       | Same as ``float(value)``           |
+        +--------------------------------+------------------------------------+
+        | Dataclasses                    | Same as :func:`dataclasses.asdict` |
+        +--------------------------------+------------------------------------+
+        | :class:`ipaddress.IPv4Address` | Same as ``str(value.exploded)``    |
+        | :class:`ipaddress.IPv6Address` |                                    |
+        +--------------------------------+------------------------------------+
+        | :class:`pathlib.Path`          | Same as ``str(value)``             |
+        +--------------------------------+------------------------------------+
+        | :class:`array.array`           | Same as :meth:`array.array.tolist` |
+        +--------------------------------+------------------------------------+
 
         """
-        if isinstance(obj, uuid.UUID):
+        if isinstance(obj, (pathlib.Path, uuid.UUID)):
             return str(obj)
         if hasattr(obj, 'isoformat'):
-            return typing.cast(type_info.DefinesIsoFormat, obj).isoformat()
+            return typing.cast(type_info.SupportsIsoFormat, obj).isoformat()
         if isinstance(obj, (bytes, bytearray, memoryview)):
             return base64.b64encode(obj).decode('ASCII')
+        if isinstance(obj, (ipaddress.IPv4Address, ipaddress.IPv6Address)):
+            return obj.exploded
+        if isinstance(obj, decimal.Decimal):
+            return float(obj)
+        if dataclasses.is_dataclass(obj):
+            return dataclasses.asdict(obj)
+        if isinstance(obj, array.array):
+            return obj.tolist()
         raise TypeError('{!r} is not JSON serializable'.format(obj))
 
 
@@ -196,6 +220,18 @@ class MsgPackTranscoder(handlers.BinaryContentHandler):
         +-----------------------------------+-------------------------------+
         | :class:`uuid.UUID`                | Converted to String           |
         +-----------------------------------+-------------------------------+
+        | :class:`decimal.Decimal`          | `float family`_               |
+        +-----------------------------------+-------------------------------+
+        | :class:`ipaddress.IPv4Address`    | `str family`_ of the exploded |
+        | :class:`ipaddress.IPv6Address`    | address form                  |
+        +-----------------------------------+-------------------------------+
+        | Dataclasses                       | `map family`_ after calling   |
+        |                                   | :func:`dataclasses.asdict`    |
+        +-----------------------------------+-------------------------------+
+        | :class:`pathlib.Path`             | Same as ``str(value)``        |
+        +-----------------------------------+-------------------------------+
+        | :class:`array.array`              | `array family`_               |
+        +-----------------------------------+-------------------------------+
 
         .. _nil byte: https://github.com/msgpack/msgpack/blob/
            0b8f5ac67cdd130f4d4d4fe6afb839b989fdb86a/spec.md#formats-nil
@@ -221,10 +257,13 @@ class MsgPackTranscoder(handlers.BinaryContentHandler):
         if datum is None:
             return datum
 
+        if isinstance(datum, decimal.Decimal):
+            datum = float(datum)
+
         if isinstance(datum, self.PACKABLE_TYPES):
             return datum
 
-        if isinstance(datum, uuid.UUID):
+        if isinstance(datum, (pathlib.Path, uuid.UUID)):
             datum = str(datum)
 
         if isinstance(datum, bytearray):
@@ -234,19 +273,25 @@ class MsgPackTranscoder(handlers.BinaryContentHandler):
             datum = datum.tobytes()
 
         if hasattr(datum, 'isoformat'):
-            datum = typing.cast(type_info.DefinesIsoFormat, datum).isoformat()
+            datum = typing.cast(type_info.SupportsIsoFormat, datum).isoformat()
+
+        if isinstance(datum, (ipaddress.IPv4Address, ipaddress.IPv6Address)):
+            datum = datum.exploded
 
         if isinstance(datum, (bytes, str)):
             return datum
 
-        if isinstance(datum, (collections.abc.Sequence, collections.abc.Set)):
-            return [self.normalize_datum(item) for item in datum]
+        if dataclasses.is_dataclass(datum):
+            datum = dataclasses.asdict(datum)
 
         if isinstance(datum, collections.abc.Mapping):
             out = {}
             for k, v in datum.items():
                 out[k] = self.normalize_datum(v)
             return out
+
+        if isinstance(datum, (collections.abc.Iterable, collections.abc.Set)):
+            return [self.normalize_datum(item) for item in datum]
 
         raise TypeError('{} is not msgpackable'.format(
             datum.__class__.__name__))
@@ -261,12 +306,13 @@ class FormUrlEncodingOptions:
     encode_sequences: bool = False
     """Encode sequence values as multiple name=value instances."""
 
-    literal_mapping: dict[typing.Literal[None, True, False],
-                          str] = dataclasses.field(default_factory=lambda: {
-                              None: '',
-                              True: 'true',
-                              False: 'false'
-                          })
+    literal_mapping: typing.Dict[typing.Union[None, bool],
+                                 str] = dataclasses.field(
+                                     default_factory=lambda: {
+                                         None: '',
+                                         True: 'true',
+                                         False: 'false'
+                                     })
     """Mapping from supported literal values to strings."""
 
     space_as_plus: bool = False
@@ -284,29 +330,40 @@ class FormUrlEncodedTranscoder:
     sequences of pairs and encodes both the name and value.  The
     following table describes how each supported type is encoded.
 
-    +----------------------------+---------------------------------------+
-    | Value / Type               | Encoding                              |
-    +============================+=======================================+
-    | character strings          | UTF-8 codepoints before percent-      |
-    |                            | encoding the resulting bytes          |
-    +----------------------------+---------------------------------------+
-    | space character            | ``%20`` or ``+``                      |
-    +----------------------------+---------------------------------------+
-    | :data:`False`              | ``false``                             |
-    +----------------------------+---------------------------------------+
-    | :data:`True`               | ``true``                              |
-    +----------------------------+---------------------------------------+
-    | :data:`None`               | the empty string                      |
-    +----------------------------+---------------------------------------+
-    | numbers                    | ``str(n)``                            |
-    +----------------------------+---------------------------------------+
-    | byte sequences             | percent-encoded bytes                 |
-    +----------------------------+---------------------------------------+
-    | :class:`uuid.UUID`         | ``str(u)``                            |
-    +----------------------------+---------------------------------------+
-    | :class:`datetime.datetime` | result of calling                     |
-    |                            | :meth:`~datetime.datetime.isoformat`  |
-    +----------------------------+---------------------------------------+
+    +--------------------------------+---------------------------------------+
+    | Value / Type                   | Encoding                              |
+    +================================+=======================================+
+    | character strings              | UTF-8 codepoints before percent-      |
+    |                                | encoding the resulting bytes          |
+    +--------------------------------+---------------------------------------+
+    | space character                | ``%20`` or ``+``                      |
+    +--------------------------------+---------------------------------------+
+    | :data:`False`                  | ``false``                             |
+    +--------------------------------+---------------------------------------+
+    | :data:`True`                   | ``true``                              |
+    +--------------------------------+---------------------------------------+
+    | :data:`None`                   | the empty string                      |
+    +--------------------------------+---------------------------------------+
+    | numbers                        | ``str(n)``                            |
+    +--------------------------------+---------------------------------------+
+    | byte sequences                 | percent-encoded bytes                 |
+    +--------------------------------+---------------------------------------+
+    | :class:`uuid.UUID`             | ``str(u)``                            |
+    +--------------------------------+---------------------------------------+
+    | :class:`datetime.datetime`     | result of calling                     |
+    |                                | :meth:`~datetime.datetime.isoformat`  |
+    +--------------------------------+---------------------------------------+
+    | Dataclasses                    | same as calling                       |
+    |                                | :func:`dataclasses.asdict` on value   |
+    +--------------------------------+---------------------------------------+
+    | :class:`ipaddress.IPv4Address` | ``str(v.exploded)``                   |
+    | :class:`ipaddress.IPv6Address` |                                       |
+    +--------------------------------+---------------------------------------+
+    | :class:`pathlib.Path`          | Same as ``str(value)``                |
+    +--------------------------------+---------------------------------------+
+    | :class:`array.array`           | Same as encoding                      |
+    |                                | :meth:`array.array.tolist`            |
+    +--------------------------------+---------------------------------------+
 
     https://url.spec.whatwg.org/#application/x-www-form-urlencoded
 
@@ -429,30 +486,41 @@ class FormUrlEncodedTranscoder:
         return dict(output)
 
     def _encode(self, datum: typing.Union[bool, None, float, int, str,
-                                          type_info.DefinesIsoFormat],
+                                          type_info.SupportsIsoFormat,
+                                          array.array[typing.Any]],
                 char_map: typing.Mapping[int, str], encoding: str) -> str:
         if isinstance(datum, str):
-            pass  # optimization: skip additional checks for strings
+            str_repr = datum
         elif (isinstance(datum, (float, int, str, uuid.UUID))
               and not isinstance(datum, bool)):
-            datum = str(datum)
+            str_repr = str(datum)
+        elif isinstance(datum, (ipaddress.IPv4Address, ipaddress.IPv6Address)):
+            str_repr = datum.exploded
+        elif isinstance(datum, array.array):
+            str_repr = str(datum.tolist())
         elif (isinstance(datum, collections.abc.Hashable)
               and datum in self.options.literal_mapping):
             # the isinstance Hashable check confuses mypy
-            datum = self.options.literal_mapping[datum]  # type: ignore
+            str_repr = self.options.literal_mapping[datum]  # type: ignore
         elif isinstance(datum, (bytearray, bytes, memoryview)):
             return ''.join(char_map[c] for c in datum)
-        elif isinstance(datum, type_info.DefinesIsoFormat):
-            datum = datum.isoformat()
+        elif isinstance(datum, type_info.SupportsIsoFormat):
+            str_repr = datum.isoformat()
+        # the following check is for namedtuples
+        elif isinstance(datum, tuple) and datum.__class__ is not tuple:
+            str_repr = str(tuple(datum))
         else:
-            datum = str(datum)
+            str_repr = str(datum)
 
-        return ''.join(char_map[c] for c in datum.encode(encoding))
+        return ''.join(char_map[c] for c in str_repr.encode(encoding))
 
     def _convert_to_tuple_sequence(
         self, value: type_info.Serializable
     ) -> typing.Iterable[typing.Tuple[typing.Any, typing.Any]]:
         tuples: typing.Iterable[typing.Tuple[typing.Any, typing.Any]]
+        if dataclasses.is_dataclass(value):
+            value = dataclasses.asdict(value)
+
         if isinstance(value, collections.abc.Mapping):
             tuples = value.items()
         else: